change the ne arguments for ggml_interpolate to int64 - more consistent with ne being int64 everywhere else

+++ include/ggml.h
@@ -1798,10 +1798,10 @@ extern "C" {
     GGML_API struct ggml_tensor * ggml_interpolate(
             struct ggml_context * ctx,
             struct ggml_tensor  * a,
-            int                   ne0,
-            int                   ne1,
-            int                   ne2,
-            int                   ne3,
+            int64_t               ne0,
+            int64_t               ne1,
+            int64_t               ne2,
+            int64_t               ne3,
             uint32_t              mode); // ggml_scale_mode [ | ggml_scale_flag...]
 
     // pad each dimension with zeros: [x, ..., x] -> [x, ..., x, 0, ..., 0]
